+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 ### Added
+
+## [0.1.11]
+### Added
 - `Verik` annotation to enable inspections by IntelliJ plugin.
 - File annotation checker issues warnings if `Verik` annotation is missing.
 - Consolidate interpreter checks to midCheck.